SIMPLE BITCOIN PREDICTION USING LSTM NETWORKS
This algorithm set a connexion to Binance and receives historical data of a crypto-currency active analysing only one variable. In this case the variable analyzed to predict is "Close price" in daily timeframe. The window to predict is one day in advantage.
Price of bitcoin is column 3 of dataset. It is possible use this algorithm with any active.